langchain/Libraries_External/v1.0.2(openAI LangchainApi 调用)/requirements.txt
langchain>=0.1.0
openai>=1.0.0
python-dotenv>=0.19.0
faiss-cpu>=1.7.4
tiktoken>=0.5.1 